fix: stabilize favorites rollout
Run runtime migrations before API startup queries so deployed builds have the latest booru site schema before handling requests.

Keep the gallery detail route cleanup tied to the stable close handler so a re-render after opening a file does not immediately close the detail panel.

Fixes #135
